CSEP576-1: First/SecondDerivImage implemented
Both functions use the simple kernel from the lecture slides to quickly
create first and second derivatives.  After convolving across that,
gaussian blur is applied to smooth things out.

Not entirely sure why the kernel for the second derivative needed to be
inverted...I'll think about that one for a bit longer.