Tweak image pixel intensity calculations
Apparently SSD and SAD needed some tweaking to get the results into
acceptable ranges.  The homework suggests we need to be able to generate
images with errors below a certain threshold, and I needed these tweaks
to be able to hit that.  I understand removing the 0.3/0.6/0.1 part for
intensity calculation, but I'm less happy with adding a sqrt for SSD,
since I don't know where that idea came from.  But someone on the forum
suggested it, and the TA seemed to concur, so who am I to argue with
wishy-washy logic like that.

The README presumes I'll use a late day on this, which I am likely to do
(so that I can finish the written assignment, booooo large papers).
I'll update it if I need to.